updated: create composables & models dashboard
