+++ features/dashboard/models/model-user.ts
@@ -1,0 +1,7 @@
+export const modelUser = (data: any[], pageSize: number, page: number) => {
+  return data?.map((x, i) => ({
+    ...x,
+    key: i,
+    no: i + 1 + pageSize * (page - 1),
+  }))
+}
