+++ features/dashboard/ui/index.vue
@@ -1,32 +1,13 @@
 <script lang="ts" setup>
-import { useQuery } from '@tanstack/vue-query'
+import useGetList from '../composables/useGetList'
 
-const fetcher = (page: Ref<number>) =>
-  fetch(
-    `https://reqres.in/api/users?page=${page.value}&per_page=3`,
-  ).then((response: any) => response.json())
+const {
+  fetchQuery,
+  dataModified,
+  handlePageChange,
+} = useGetList()
 
-const page = ref(1)
-
-const { isLoading, isError, data, error } = useQuery({
-  queryKey: ['users', page],
-  queryFn: () => fetcher(page),
-  keepPreviousData: true,
-})
-
-const dataModified = computed(() => {
-  if (data.value) {
-    return data.value.data.map((x: any, i: any) => ({
-      ...x,
-      key: i,
-      no: i + 1 + data.value.per_page * (page.value - 1),
-    }))
-  }
-})
-
-const handlePageChange = (newPage: any) => {
-  page.value = newPage
-}
+const { isLoading, data } = fetchQuery
 
 const columns = [
   {
